reorganize date formatting composable
