+++ src/composables/index.ts
@@ -1,2 +1,5 @@
 export * from "./useFormat";
 export * from "./useWebExtStorage";
+export * from "./useNumericMask";
+export * from "./useDateFormat";
+export * from "./useProgrammaticDialog";
